+++ CHANGELOG.rst
@@ -66,6 +66,8 @@ Important changes (that users especially need to know) are marked by '**[!]**.'
 
 * Add toc_depth option to wkhtmltopdf converter.
 
+* Add ftype option
+
 **Remove:**
 
 * Remove 'support' for ebook-convert. Now converters are only one of the three